% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afr.R
\docType{data}
\name{afr}
\alias{afr}
\title{African ancestry LD block breakpoints}
\format{An object of class \code{data.frame} with 2583 rows and 3 columns.}
\usage{
data(afr)
}
\description{
African ancestry LD block breakpoints
}
\references{
Berisa T et al. (2016). Approximately independent linkage disequilibrium blocks in human populations. \emph{Bioinformatics}.
}
\keyword{datasets}
