% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn.R
\docType{data}
\name{asn}
\alias{asn}
\title{Asian ancestry LD block breakpoints}
\format{An object of class \code{data.frame} with 1445 rows and 3 columns.}
\usage{
data(asn)
}
\description{
Asian ancestry LD block breakpoints
}
\references{
Berisa T et al. (2016). Approximately independent linkage disequilibrium blocks in human populations. \emph{Bioinformatics}.
}
\keyword{datasets}
