% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\docType{data}
\name{eur}
\alias{eur}
\title{European ancestry LD block breakpoints}
\format{An object of class \code{data.frame} with 1703 rows and 3 columns.}
\usage{
data(eur)
}
\description{
European ancestry LD block breakpoints
}
\references{
Berisa T et al. (2016). Approximately independent linkage disequilibrium blocks in human populations. \emph{Bioinformatics}.
}
\keyword{datasets}
